requirements.txt
alembic==1.6.4
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
distro==1.5.0
dnspython==2.1.0
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.1.2
Flask==2.0.1
flask-mongoengine==1.0.0
flask-swagger==0.2.14
Flask-WTF==0.15.0
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
mongoengine==0.23.1
paramiko==2.7.2
pycparser==2.20
pymongo==3.11.4
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-editor==1.0.4
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
texttable==1.6.3
urllib3==1.26.4
websocket-client==0.59.0
Werkzeug==2.0.1
WTForms==2.3.3
